+++ job-tracker-core/src/main/java/org/limbo/flowjob/tracker/core/job/context/JobRecordRepository.java
@@ -19,12 +19,12 @@ public interface JobRecordRepository {
     /**
      * 获取
      */
-    JobRecord get(String planId, Long planRecordId, Long planInstanceId, String jobId);
+    JobRecord get(String planId, Long planRecordId, Integer planInstanceId, String jobId);
 
-    void executing(String planId, Long planRecordId, Long planInstanceId, String jobId);
+    void executing(String planId, Long planRecordId, Integer planInstanceId, String jobId);
 
-    void end(String planId, Long planRecordId, Long planInstanceId, String jobId, JobScheduleStatus state);
+    void end(String planId, Long planRecordId, Integer planInstanceId, String jobId, JobScheduleStatus state);
 
-    List<JobRecord> getRecords(String planId, Long planRecordId, Long planInstanceId, List<String> jobIds);
+    List<JobRecord> getRecords(String planId, Long planRecordId, Integer planInstanceId, List<String> jobIds);
 
 }
