+++ job-tracker-dao/src/main/java/org/limbo/flowjob/tracker/dao/mybatis/JobInstanceMapper.java
@@ -14,7 +14,7 @@ public interface JobInstanceMapper extends BaseMapper<JobInstancePO> {
     @Select("select job_instance_id from flowjob_job_instance " +
             "where plan_id = #{planId} and plan_record_id = #{planRecordId} and plan_instance_id = #{planInstanceId} and job_id = #{jobId} " +
             "order by plan_instance_id desc limit 1 for update")
-    Long getRecentlyIdForUpdate(@Param("planId") String planId, @Param("planRecordId") Long planRecordId,
-                                @Param("planInstanceId") Long planInstanceId, @Param("jobId") String jobId);
+    Integer getRecentlyIdForUpdate(@Param("planId") String planId, @Param("planRecordId") Long planRecordId,
+                                @Param("planInstanceId") Integer planInstanceId, @Param("jobId") String jobId);
 
 }
