+++ job-tracker-dao/src/main/java/org/limbo/flowjob/tracker/dao/mybatis/PlanInstanceMapper.java
@@ -12,5 +12,5 @@ import org.limbo.flowjob.tracker.dao.po.PlanInstancePO;
 public interface PlanInstanceMapper extends BaseMapper<PlanInstancePO> {
 
     @Select("select plan_instance_id from flowjob_plan_instance where plan_id = #{planId} and plan_record_id = #{planRecordId} order by plan_instance_id desc limit 1 for update")
-    Long getRecentlyIdForUpdate(@Param("planId") String planId, @Param("planRecordId") Long planRecordId);
+    Integer getRecentlyIdForUpdate(@Param("planId") String planId, @Param("planRecordId") Long planRecordId);
 }
